+++ com.unity.render-pipelines.high-definition/Runtime/Debug/LightingDebug.cs
@@ -319,6 +319,8 @@ namespace UnityEngine.Rendering.HighDefinition
         public float                debugExposure = 0.0f;
         /// <summary>Whether to show tonemap curve in the histogram debug view or not.</summary>
         public bool                 showTonemapCurveAlongHistogramView = true;
+        /// <summary>Whether to center the histogram debug view around the middle-grey point or not.</summary>
+        public bool                 centerHistogramAroundMiddleGrey = false;
 
         /// <summary>Display the light cookies atlas.</summary>
         public bool                 displayCookieAtlas = false;